+++ tests/test___main__.py
@@ -21,27 +21,38 @@ import datetime
 import os
 import subprocess
 
+import pytest
+
+
 # pylint: disable=redefined-outer-name, invalid-name
 # pylint: disable=missing-class-docstring, missing-function-docstring
 # pylint: disable=too-many-arguments, unused-argument, too-few-public-methods
 
 
-class Test_CLI_exit_code:
-    def test_when_successful_then_exit_code_is_zero(self, yaml_path_all, tmp_path):
-        assert execute_cli("-o", tmp_path, "-g", "python", yaml_path_all).returncode == 0
+package_path = {"python": ("__init__.py",), "cpp": ("include", "p_p", "p_p.hpp")}
 
-    def test_when_invalid_input_path_then_exit_code_is_one(self, yaml_path_all, tmp_path):
-        assert execute_cli("-o", tmp_path, "-g", "python", "invalid").returncode == 1
+
+class Test_CLI_exit_code:
+    @pytest.mark.parametrize("generator", ("python", "cpp"))
+    def test_when_successful_then_exit_code_is_zero(self, yaml_path_all, tmp_path, generator):
+        assert execute_cli("-o", tmp_path, "-g", generator, yaml_path_all).returncode == 0
+
+    @pytest.mark.parametrize("generator", ("python", "cpp"))
+    def test_when_invalid_input_path_then_exit_code_is_one(
+        self, yaml_path_all, tmp_path, generator
+    ):
+        assert execute_cli("-o", tmp_path, "-g", generator, "invalid").returncode == 1
 
     def test_when_invalid_args_set_then_exit_code_is_two(self):
         assert execute_cli("--invalid-arg").returncode == 2
 
-    def test_when_o_flag_not_set_error_is_output(self, yaml_path_all, tmp_path):
+    @pytest.mark.parametrize("generator", ("python", "cpp"))
+    def test_when_o_flag_not_set_error_is_output(self, yaml_path_all, tmp_path, generator):
         # stderr also outputs all the usage information which we don't
         # care about for the purposes of this test.
         assert (
             "error: the following arguments are required: -o/--output-dir"
-            in execute_cli("-g", "python", yaml_path_all).stderr
+            in execute_cli("-g", generator, yaml_path_all).stderr
         )
 
     def test_when_g_flag_not_set_error_is_output(self, yaml_path_all, tmp_path):
@@ -53,136 +64,173 @@ class Test_CLI_exit_code:
         )
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_output:
-    def test_when_verbose_off_then_stdout_is_empty(self, yaml_path_minimal, tmp_path):
-        assert execute_cli("--generator", "python", "-o", tmp_path, yaml_path_minimal).stdout == ""
+    def test_when_verbose_off_then_stdout_is_empty(self, yaml_path_minimal, tmp_path, generator):
+        assert (
+            execute_cli("--generator", generator, "-o", tmp_path, yaml_path_minimal).stdout == ""
+        )
 
     def test_when_verbose_on_then_created_paths_written_to_stdout(
-        self, yaml_path_minimal, creations_minimal_python, tmp_path
+        self, yaml_path_minimal, creations_minimal_by_generator, tmp_path, generator
     ):
-        expected = [os.path.join(tmp_path, path) for path in creations_minimal_python]
+        expected = [
+            os.path.join(tmp_path, path) for path in creations_minimal_by_generator[generator]
+        ]
+
         actual = execute_cli(
-            "--generator", "python", "-o", tmp_path, yaml_path_minimal, "-v"
+            "--generator", generator, "-o", tmp_path, yaml_path_minimal, "-v"
         ).stdout.splitlines()
 
         assert actual == expected
 
     def test_when_verbose_and_logging_on_then_created_paths_written_to_stdout(
-        self, yaml_path_minimal, creations_minimal_python, tmp_path
+        self, yaml_path_minimal, creations_minimal_by_generator, tmp_path, generator
     ):
-        expected = [os.path.join(tmp_path, path) for path in creations_minimal_python]
+        expected = [
+            os.path.join(tmp_path, path) for path in creations_minimal_by_generator[generator]
+        ]
         actual = execute_cli(
-            "--generator", "python", "-o", tmp_path, yaml_path_minimal, "-v", "-l", "DEBUG"
+            "--generator",
+            generator,
+            "-o",
+            tmp_path,
+            yaml_path_minimal,
+            "-v",
+            "-l",
+            "DEBUG",
         ).stdout.splitlines()
 
         assert actual == expected
 
     def test_when_logging_not_set_then_INFO_messages_not_written_to_stderr(
-        self, yaml_path_minimal, tmp_path
+        self, yaml_path_minimal, tmp_path, generator
     ):
         assert (
             execute_cli(
-                "--generator", "python", "-o", tmp_path, yaml_path_minimal
+                "--generator", generator, "-o", tmp_path, yaml_path_minimal
             ).stderr.splitlines()
             == []
         )
 
     def test_when_logging_set_to_INFO_then_INFO_messages_written_to_stderr(
-        self, yaml_path_minimal, tmp_path
+        self, yaml_path_minimal, tmp_path, generator
     ):
         assert (
             "INFO:"
             in execute_cli(
-                "--generator", "python", "-o", tmp_path, "-l", "INFO", yaml_path_minimal
+                "--generator",
+                generator,
+                "-o",
+                tmp_path,
+                "-l",
+                "INFO",
+                yaml_path_minimal,
             ).stderr
         )
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_args_dry_run:
     def test_when_not_set_then_code_is_generated_to_expected_path(
-        self, yaml_path_minimal, tmp_path
+        self, yaml_path_minimal, tmp_path, generator
     ):
-        execute_cli("--generator", "python", "-o", tmp_path, yaml_path_minimal)
+        execute_cli("--generator", generator, "-o", tmp_path, yaml_path_minimal)
 
         assert os.path.isdir(os.path.join(tmp_path, "p_p"))
 
-    def test_when_d_set_then_code_is_not_generated(self, yaml_path_minimal, tmp_path):
-        execute_cli("--generator", "python", "-d", "-o", tmp_path, yaml_path_minimal)
+    def test_when_d_set_then_code_is_not_generated(self, yaml_path_minimal, tmp_path, generator):
+        execute_cli("--generator", generator, "-d", "-o", tmp_path, yaml_path_minimal)
 
         assert not os.path.isdir(os.path.join(tmp_path, "p_p"))
 
-    def test_when_dryrun_set_then_code_is_not_generated(self, yaml_path_minimal, tmp_path):
-        execute_cli("--generator", "python", "--dry-run", "-o", tmp_path, yaml_path_minimal)
+    def test_when_dryrun_set_then_code_is_not_generated(
+        self, yaml_path_minimal, tmp_path, generator
+    ):
+        execute_cli("--generator", generator, "--dry-run", "-o", tmp_path, yaml_path_minimal)
 
         assert not os.path.isdir(os.path.join(tmp_path, "p_p"))
 
-    def test_when_set_then_exit_code_is_zero(self, yaml_path_minimal, tmp_path):
+    def test_when_set_then_exit_code_is_zero(self, yaml_path_minimal, tmp_path, generator):
         assert (
             execute_cli(
-                "--generator", "python", "-d", "-o", tmp_path, yaml_path_minimal
+                "--generator", generator, "-d", "-o", tmp_path, yaml_path_minimal
             ).returncode
             == 0
         )
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_args_output_dir:
-    def test_when_not_set_then_exit_code_is_two(self, yaml_path_minimal):
-        assert execute_cli("--generator", "python", yaml_path_minimal).returncode == 2
+    def test_when_not_set_then_exit_code_is_two(self, yaml_path_minimal, generator):
+        assert execute_cli("--generator", generator, yaml_path_minimal).returncode == 2
 
-    def test_when_o_set_then_code_is_generated_to_expected_path(self, yaml_path_minimal, tmp_path):
-        execute_cli("--generator", "python", "-o", tmp_path, yaml_path_minimal)
+    def test_when_o_set_then_code_is_generated_to_expected_path(
+        self, yaml_path_minimal, tmp_path, generator
+    ):
+        execute_cli("--generator", generator, "-o", tmp_path, yaml_path_minimal)
 
         assert os.path.isdir(os.path.join(tmp_path, "p_p"))
 
     def test_when_outputdir_set_then_code_is_generated_to_expected_path(
-        self, yaml_path_minimal, tmp_path
+        self, yaml_path_minimal, tmp_path, generator
     ):
-        execute_cli("--generator", "python", "--output-dir", tmp_path, yaml_path_minimal)
+        execute_cli("--generator", generator, "--output-dir", tmp_path, yaml_path_minimal)
 
         assert os.path.isdir(os.path.join(tmp_path, "p_p"))
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_args_copyright_owner:
-    def test_when_not_set_then_no_copyright_added(self, tmp_path, yaml_path_minimal):
-        execute_cli("--generator", "python", "-o", tmp_path, yaml_path_minimal)
+    def test_when_not_set_then_no_copyright_added(self, tmp_path, yaml_path_minimal, generator):
+        execute_cli("--generator", generator, "-o", tmp_path, yaml_path_minimal)
 
-        assert "SPDX-License-Identifier:" not in file_contents(tmp_path, "p_p", "__init__.py")
+        assert "SPDX-License-Identifier:" not in file_contents(
+            tmp_path, "p_p", *package_path[generator]
+        )
 
-    def test_when_set_then_copyright_added_with_owner(self, tmp_path, yaml_path_minimal):
+    def test_when_set_then_copyright_added_with_owner(
+        self, tmp_path, yaml_path_minimal, generator
+    ):
         execute_cli(
             "--generator",
-            "python",
+            generator,
             "--copyright-owner",
             "An Owner",
             "-o",
             tmp_path,
             yaml_path_minimal,
         )
-        contents = file_contents(tmp_path, "p_p", "__init__.py")
+        contents = file_contents(tmp_path, "p_p", *package_path[generator])
 
         assert "SPDX-License-Identifier:" in contents
         assert "An Owner" in contents
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_args_copyright_date:
-    def test_when_not_set_then_copyright_date_is_this_year(self, tmp_path, yaml_path_minimal):
+    def test_when_not_set_then_copyright_date_is_this_year(
+        self, tmp_path, yaml_path_minimal, generator
+    ):
         execute_cli(
             "--generator",
-            "python",
+            generator,
             "--copyright-owner",
             "An Owner",
             "-o",
             tmp_path,
             yaml_path_minimal,
         )
-        contents = file_contents(tmp_path, "p_p", "__init__.py")
+        contents = file_contents(tmp_path, "p_p", *package_path[generator])
 
         assert f"Copyright {datetime.date.today().year} An Owner" in contents
 
-    def test_when_set_then_copyright_date_is_as_specified(self, tmp_path, yaml_path_minimal):
+    def test_when_set_then_copyright_date_is_as_specified(
+        self, tmp_path, yaml_path_minimal, generator
+    ):
         execute_cli(
             "--generator",
-            "python",
+            generator,
             "--copyright-owner",
             "An Owner",
             "--copyright-date",
@@ -191,32 +239,33 @@ class Test_CLI_args_copyright_date:
             tmp_path,
             yaml_path_minimal,
         )
-        contents = file_contents(tmp_path, "p_p", "__init__.py")
+        contents = file_contents(tmp_path, "p_p", *package_path[generator])
 
         assert "Copyright 2010-2022 An Owner" in contents
 
 
+@pytest.mark.parametrize("generator", ("python", "cpp"))
 class Test_CLI_args_spdxLicenseIdentifier:
-    def test_when_not_set_then_apache_two_used(self, tmp_path, yaml_path_minimal):
+    def test_when_not_set_then_apache_two_used(self, tmp_path, yaml_path_minimal, generator):
         execute_cli(
             "--generator",
-            "python",
+            generator,
             "--copyright-owner",
             "An Owner",
             "-o",
             tmp_path,
             yaml_path_minimal,
         )
-        contents = file_contents(tmp_path, "p_p", "__init__.py")
+        contents = file_contents(tmp_path, "p_p", *package_path[generator])
 
         assert "SPDX-License-Identifier: Apache-2.0" in contents
 
     def test_when_set_then_license_added_with_specified_identifier(
-        self, tmp_path, yaml_path_minimal
+        self, tmp_path, yaml_path_minimal, generator
     ):
         execute_cli(
             "--generator",
-            "python",
+            generator,
             "--spdx-license-identifier",
             "Unlicense",
             "--copyright-owner",
@@ -225,7 +274,7 @@ class Test_CLI_args_spdxLicenseIdentifier:
             tmp_path,
             yaml_path_minimal,
         )
-        contents = file_contents(tmp_path, "p_p", "__init__.py")
+        contents = file_contents(tmp_path, "p_p", *package_path[generator])
 
         assert "SPDX-License-Identifier: Unlicense" in contents
 
@@ -238,9 +287,12 @@ class Test_CLI_args_help:
         assert "usage: openassetio-traitgen" in execute_cli("--help").stdout
 
 
-class Test_CLI_args_python:
-    def test_when_python_set_then_python_is_generated(self, tmp_path, yaml_path_minimal):
-        execute_cli("--generator", "python", "-o", tmp_path, yaml_path_minimal)
+class Test_CLI_args:
+    @pytest.mark.parametrize("generator", ("python", "cpp"))
+    def test_when_generator_set_then_correct_language_is_generated(
+        self, tmp_path, yaml_path_minimal, generator
+    ):
+        execute_cli("--generator", generator, "-o", tmp_path, yaml_path_minimal)
 
         assert os.path.isdir(os.path.join(tmp_path, "p_p"))