+++ RELEASE_NOTES.md
@@ -4,10 +4,19 @@ Release Notes
 v1.0.0-alpha.X
 --------------
 
+### New Features
+
+- Added C++ trait and specification class generation, where the
+  generated source tree consists of a header-only package, with a header
+  file per class along with hoisting headers for convenience, broadly
+  mirroring the source tree of the Python generator.
+  [#11](https://github.com/OpenAssetIO/OpenAssetIO-TraitGen/issues/11)
+
 ### Bug fixes
 
 - Fixed line breaks to no longer be platform-specific, and instead
   conform to Unix-style `\n`.
+
 
 v1.0.0-alpha.4
 --------------